迁移position_utils.py和note2display
- position_utils.py迁移到org_utils（见上一个commit）
- note2dis从views迁移到note_utils
