+++ app/views.py
@@ -39,6 +39,7 @@ from app.notification_utils import(
     notification_create,
     bulk_notification_create,
     notification_status_change,
+    notification2Display,
 )
 from app.QA_utils import (
     QA2Display,
@@ -1773,51 +1774,6 @@ def apply_position(request, oid=None):
     )
     return redirect("/notifications/")
 '''
-
-
-
-
-@log.except_captured(source='views[notification2Display]')
-def notification2Display(notification_set):
-    lis = []
-    sender_userids = notification_set.values_list('sender_id', flat=True)
-    sender_persons = NaturalPerson.objects.filter(person_id__in=sender_userids).values_list('person_id', 'name')
-    sender_persons = {userid: name for userid, name in sender_persons}
-    sender_orgs = Organization.objects.filter(organization_id__in=sender_userids).values_list('organization_id', 'oname')
-    sender_orgs = {userid: name for userid, name in sender_orgs}
-    # 储存这个列表中所有record的元气值的和
-    for notification in notification_set:
-        note_display = {}
-
-        # id
-        note_display["id"] = notification.id
-
-        # 时间
-        note_display["start_time"] = notification.start_time.strftime("%Y-%m-%d %H:%M")
-        if notification.finish_time is not None:
-            note_display["finish_time"] = notification.finish_time.strftime("%Y-%m-%d %H:%M")
-
-        # 留言
-        note_display["content"] = notification.content
-
-        # 状态
-        note_display["status"] = notification.get_status_display()
-        note_display["URL"] = notification.URL
-        note_display["type"] = notification.get_typename_display()
-        note_display["title"] = notification.get_title_display()
-
-
-        _, user_type, _ = utils.check_user_type(notification.sender)
-        if user_type == "Organization":
-            note_display["sender"] = sender_orgs.get(
-                notification.sender_id
-            ) if not notification.anonymous_flag else "匿名者"
-        else:
-            note_display["sender"] = sender_persons.get(
-                notification.sender_id
-            ) if not notification.anonymous_flag else "匿名者"
-        lis.append(note_display)
-    return lis
 
 
 @login_required(redirect_field_name="origin")
